# backend/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Worship Room API
  description: |
    Backend API for Worship Room.

    Phase 1: Proxy endpoints for upstream APIs (Gemini, Google Places, FCBH)
    that previously had keys exposed in the frontend bundle. Subsequent
    Forums Wave specs add application endpoints (auth, posts, comments, etc.).

    See `_specs/ai-proxy-foundation.md` for proxy architecture details.
  version: 1.0.0
  license:
    name: Proprietary
    identifier: proprietary
  contact:
    name: Eric Champlin
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.worshiproom.com
    description: Production
...
paths:
...
  /api/v1/posts:
    get:
      tags: [Posts]
      summary: List posts (paginated public feed)
      description: |
        Paginated feed of posts. Optional auth via Bearer JWT — anonymous
        viewers see only public posts; authenticated viewers see public +
        relevant friends-visibility + their own posts. Muted authors
        filtered out for authenticated viewers.

        Returns 200 with empty `data: []` when no posts match (NOT 404).
      operationId: listPosts
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [health, mental-health, family, work, grief, gratitude, praise, relationships, other, discussion]
        - name: postType
          in: query
          required: false
          schema:
            type: string
            enum: [prayer_request, testimony, question, discussion, encouragement]
        - name: qotdId
          in: query
          required: false
          schema: { type: string }
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [bumped, recent, answered]
            default: bumped
      responses:
        '200':
          description: Paginated posts (possibly empty)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Posts]
      summary: Create a new post
      description: |
        Authenticated post creation. Crisis-keyword detection runs server-side;
        flagged posts are STILL created (the `crisisFlag` field is true on the
        response, and a `crisisResources` block sits at the response root).

        Rate limit: 5 posts per 24h per user. Returns 429 RATE_LIMITED with
        `Retry-After` header when exceeded.

        Idempotency: clients may pass `Idempotency-Key` header to deduplicate
        retried POSTs. Same key + same body → returns the cached response.
        Same key + different body → returns 422 IDEMPOTENCY_KEY_MISMATCH.
      operationId: createPost
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string, maxLength: 100 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePostRequest' }
      responses:
        '201':
          description: Post created (with optional crisisResources block)
          headers:
            Location:
              schema: { type: string }
              description: URL of the newly created post.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatePostResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: Idempotency key reused with a different request body
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProxyError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/posts/{id}:
    get:
      tags: [Posts]
      summary: Get a single post by id
      description: |
        Returns a single post if visible to the caller. Mute filter is NOT
        applied here (Spec 3.3 Divergence 1) — direct ID lookups bypass
        discovery filters so notification/share-link paths still work.

        Returns 404 POST_NOT_FOUND if the post does not exist, is
        soft-deleted, or fails the visibility predicate (anti-enumeration —
        no differential between "forbidden" and "missing").
      operationId: getPostById
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Post detail
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { $ref: '#/components/schemas/PostDto' }
                  meta:
                    type: object
                    required: [requestId]
                    properties:
                      requestId: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    patch:
      tags: [Posts]
      summary: Update a post
      description: |
        Author-ownership-gated edit. Most fields editable only within the 5-minute
        window from createdAt; isAnswered and answeredText are window-exempt.
        Visibility downgrade always allowed; visibility upgrade requires window.

        postType and isAnonymous are immutable post-create. `crisisFlag` is
        server-managed and rejected if present in the body. Once set, crisis_flag
        is never cleared by an author edit (only moderators clear it).
      operationId: updatePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePostRequest' }
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { $ref: '#/components/schemas/PostDto' }
                  meta:
                    type: object
                    required: [requestId]
                    properties:
                      requestId: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Not author and not admin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProxyError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Edit window expired (or visibility upgrade outside window)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProxyError' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [Posts]
      summary: Soft-delete a post
      description: |
        Author-ownership-gated soft-delete. Idempotent — a second DELETE on an
        already-deleted post returns 204 (mission accomplished). Content is NOT
        replaced; user_id stays.

        Account-deletion-driven content replacement is Spec 10.11's concern;
        this endpoint only sets is_deleted=true, deleted_at=NOW(), updated_at=NOW().
        Comments are NOT cascade-deleted (cascade fires only on hard delete).
      operationId: deletePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Post soft-deleted (or already deleted)
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Not author and not admin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProxyError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/users/{username}/posts:
    get:
      tags: [Posts]
      summary: List posts by author
      description: |
        Lists posts authored by the given user. {username} accepts either a
        kebab-case identifier (first_name + '-' + last_name, lowercased)
        or a literal UUID. Visibility predicate + mute filter applied.

        Returns 200 with empty `data: []` if the username does not resolve
        (NOT 404 — anti-enumeration).

        Phase 8.1 will replace kebab-case derivation with users.username
        lookup; the URL surface is stable.
      operationId: listAuthorPosts
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [bumped, recent, answered]
            default: recent
      responses:
        '200':
          description: Paginated posts authored by the resolved user (possibly empty)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/posts/{id}/comments:
    get:
      tags: [Posts]
      summary: List comments for a post
      description: |
        Paginated comments for a post. Top-level comments only (parent_comment_id
        IS NULL); replies are nested under each parent in the `replies` array
        (always present, possibly empty during Phase 3 since no UI creates
        replies yet — Phase 4 enables that).

        Optional auth via Bearer JWT — anonymous viewers can read comments on
        publicly-visible posts. Pre-checks parent post visibility via the
        canonical visibility predicate (Spec 3.3). Returns 404 POST_NOT_FOUND
        when the post does not exist, is soft-deleted, or fails the visibility
        predicate (anti-enumeration).

        Soft-deleted comments and comments with moderation_status hidden/removed
        are excluded.
      operationId: listPostComments
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: Paginated top-level comments with replies nested
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/users/me/reactions:
    get:
      tags: [Posts]
      summary: Viewer's reactions across all posts (flat map)
      description: |
        Returns a flat map keyed by post UUID, where each value is a
        `{isPraying, isBookmarked}` pair. Posts where the viewer has neither
        reacted nor bookmarked are NOT in the map (frontend treats absence
        as both-false). Map iteration order is not guaranteed.

        Candle reactions (Phase 6.6 Answered Wall) are NOT included in this
        response; that endpoint will extend the shape with an `isCandled` field.

        No pagination — the response is bounded by the viewer's actual
        engagement count. Per-user data only; reactions and bookmarks made
        by other users are not visible.

        No visibility filtering — the response reflects the viewer's own
        engagement history. Frontend joins this map with the visible-post-list
        from `GET /api/v1/posts` at render time, naturally filtering posts
        the viewer can no longer see.
      operationId: getMyReactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Reactions map
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { $ref: '#/components/schemas/ReactionsResponse' }
                  meta:
                    type: object
                    required: [requestId]
                    properties:
                      requestId: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/users/me/bookmarks:
    get:
      tags: [Posts]
      summary: Viewer's bookmarked posts (paginated)
      description: |
        Paginated list of posts the viewer has bookmarked. Returns full
        PostDto rows (not just bookmark records). Same envelope shape as
        `GET /api/v1/posts`.

        Sort order is `post_bookmarks.created_at DESC` (most-recently-bookmarked
        first), which is distinct from `posts.created_at` — bookmarking an old
        post today places it at the top of this response.

        Visibility predicate applied: bookmarks for posts that are no longer
        visible to the viewer (now-private/deleted/hidden, or whose authors
        the viewer has since muted or been unfriended by) do not appear.
        Bookmarks the viewer made on their OWN posts always appear regardless
        of visibility (they explicitly bookmarked their own thing).
      operationId: getMyBookmarks
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: Paginated bookmarked posts (possibly empty)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
components:
  schemas:
...
    ProxyError:
      description: Standard error-response shape.
      type: object
      required: [code, message, requestId, timestamp]
      properties:
        code:
          type: string
          example: RATE_LIMITED
          description: Stable error code that the frontend branches on.
        message:
          type: string
          description: User-safe error message (no stack traces, no upstream details).
        requestId:
          type: string
          description: Same request ID that's set on the X-Request-Id response header.
        timestamp:
          type: string
          format: date-time
...
    AuthorDto:
      description: |
        Nested author block in PostDto. The `id` field is null for
        anonymous posts (and is rendered as `"id": null` rather than
        being omitted — frontend code distinguishes "field absent" from
        "field present-with-null").
      type: object
      required: [id, displayName, avatarUrl]
      properties:
        id:
          type: string
          format: uuid
          nullable: true
          description: Author user UUID; null for anonymous posts.
        displayName:
          type: string
          description: |
            Resolved display name per the user's display_name_preference.
            Always "Anonymous" for anonymous posts.
        avatarUrl:
          type: string
          nullable: true
          description: User avatar URL; null for anonymous posts or users with no avatar.
    PostDto:
      description: |
        Outbound post representation. Excludes `is_deleted`/`deleted_at`
        (filtered upstream) and `report_count` (moderation-internal).
      type: object
      required:
        - id
        - postType
        - content
        - isAnonymous
        - visibility
        - isAnswered
        - moderationStatus
        - crisisFlag
        - prayingCount
        - candleCount
        - commentCount
        - bookmarkCount
        - createdAt
        - updatedAt
        - lastActivityAt
        - author
      properties:
        id: { type: string, format: uuid }
        postType:
          type: string
          enum: [prayer_request, testimony, question, discussion, encouragement]
        content: { type: string }
        category:
          type: string
          nullable: true
          enum: [health, mental-health, family, work, grief, gratitude, praise, relationships, other, discussion]
        isAnonymous: { type: boolean }
        challengeId: { type: string, nullable: true }
        qotdId: { type: string, nullable: true }
        scriptureReference: { type: string, nullable: true }
        scriptureText: { type: string, nullable: true }
        visibility:
          type: string
          enum: [public, friends, private]
        isAnswered: { type: boolean }
        answeredText: { type: string, nullable: true }
        answeredAt: { type: string, format: date-time, nullable: true }
        moderationStatus:
          type: string
          enum: [approved, flagged, hidden, removed]
        crisisFlag: { type: boolean }
        prayingCount: { type: integer, minimum: 0 }
        candleCount: { type: integer, minimum: 0 }
        commentCount: { type: integer, minimum: 0 }
        bookmarkCount: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lastActivityAt: { type: string, format: date-time }
        author: { $ref: '#/components/schemas/AuthorDto' }
    PostListMeta:
      description: Pagination metadata for the post list endpoints.
      type: object
      required:
        - page
        - limit
        - totalCount
        - totalPages
        - hasNextPage
        - hasPrevPage
        - requestId
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 50 }
        totalCount: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        hasNextPage: { type: boolean }
        hasPrevPage: { type: boolean }
        requestId: { type: string }
    PostListResponse:
      description: Paginated post-list response envelope.
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/PostDto' }
        meta: { $ref: '#/components/schemas/PostListMeta' }
    CreatePostRequest:
      description: Request body for POST /api/v1/posts (Spec 3.5).
      type: object
      required: [postType, content]
      properties:
        postType:
          type: string
          enum: [prayer_request, testimony, question, discussion, encouragement]
        content:
          type: string
          maxLength: 2000
        category:
          type: string
          enum: [health, mental-health, family, work, grief, gratitude, praise, relationships, other, discussion]
          nullable: true
        isAnonymous:
          type: boolean
          default: false
        visibility:
          type: string
          enum: [public, friends, private]
          default: public
        challengeId:
          type: string
          maxLength: 50
          nullable: true
        qotdId:
          type: string
          maxLength: 50
          nullable: true
        scriptureReference:
          type: string
          maxLength: 100
          nullable: true
        scriptureText:
          type: string
          maxLength: 2000
          nullable: true
    UpdatePostRequest:
      description: Request body for PATCH /api/v1/posts/{id} (Spec 3.5). All fields optional.
      type: object
      properties:
        content: { type: string, maxLength: 2000 }
        category:
          type: string
          enum: [health, mental-health, family, work, grief, gratitude, praise, relationships, other, discussion]
        visibility:
          type: string
          enum: [public, friends, private]
        isAnswered: { type: boolean }
        answeredText: { type: string, maxLength: 2000 }
        challengeId: { type: string, maxLength: 50 }
        qotdId: { type: string, maxLength: 50 }
        scriptureReference: { type: string, maxLength: 100 }
        scriptureText: { type: string, maxLength: 2000 }
    CreatePostResponse:
      description: |
        Response envelope for POST /api/v1/posts (Spec 3.5). When the post tripped
        the crisis-keyword detector, the optional `crisisResources` block sits at
        the response root parallel to `data`.
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/PostDto' }
        crisisResources:
          $ref: '#/components/schemas/CrisisResourcesBlock'
        meta:
          type: object
          required: [requestId]
          properties:
            requestId: { type: string }
    CrisisResourcesBlock:
      description: Crisis-resource payload returned when a post is flagged.
      type: object
      required: [message, resources]
      properties:
        message: { type: string }
        resources:
          type: array
          items: { $ref: '#/components/schemas/CrisisResource' }
    CrisisResource:
      description: Single crisis-resource entry. Exactly one of phone/text is non-null.
      type: object
      required: [name, link]
      properties:
        name: { type: string }
        phone: { type: string, nullable: true }
        text: { type: string, nullable: true }
        link: { type: string, format: uri }
...
    CommentListResponse:
      description: Paginated comment-list response envelope.
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/CommentDto' }
        meta: { $ref: '#/components/schemas/PostListMeta' }
    PerPostReaction:
      description: Per-post engagement flags inside the reactions map.
      type: object
      required: [isPraying, isBookmarked]
      properties:
        isPraying: { type: boolean }
        isBookmarked: { type: boolean }
    ReactionsResponse:
      description: |
        Flat map keyed by post UUID. Posts where the viewer has neither
        reacted nor bookmarked are not in the map. Map iteration order is
        not guaranteed.
      type: object
      required: [reactions]
      properties:
        reactions:
          type: object
          additionalProperties: { $ref: '#/components/schemas/PerPostReaction' }
          description: Map of post UUID strings to PerPostReaction values.
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProxyError' }
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds until the rate limit window resets.
        X-RateLimit-Limit:
          schema: { type: integer }
        X-RateLimit-Remaining:
          schema: { type: integer }
        X-RateLimit-Reset:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProxyError' }
...
    NotFound:
      description: Resource not found (upstream 404, e.g., chapter with no audio)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProxyError' }
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProxyError' }
    Unauthorized:
      description: |
        Missing, malformed, or expired Bearer token; or token references a
        user that no longer exists. Frontend should clear the in-memory JWT
        and route to login.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProxyError' }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256-signed JWT issued by POST /api/v1/auth/login. Send as
        Authorization: Bearer <token>. Token is short-lived (1 hour);
        on 401 TOKEN_EXPIRED, the frontend re-logs in via /auth/login.